Improve a7 layout (#33)
* Improve boarder, size and header color

* show notes only if not to match weapons need space

* better type and category position

+++ web/src/main/java/de/twonirwana/infinity/WebApp.java
@@ -307,10 +307,13 @@ public class WebApp {
                     if (!file.toFile().isDirectory()) {
                         Path targetPath = targetDir.resolve(file.getFileName());
                         if (onlyCopy) {
-                            Files.copy(file, targetPath, StandardCopyOption.REPLACE_EXISTING);
+                            if (!targetPath.toFile().exists()) {
+                                Files.copy(file, targetPath, StandardCopyOption.REPLACE_EXISTING);
+                            }
                         } else {
                             Files.move(file, targetPath, StandardCopyOption.REPLACE_EXISTING);
                         }
+
                         count++;
                     }
                 }
